chore: move `isLive` into it's own file
