+++ src/site/_collections/pages.js
@@ -27,7 +27,7 @@ const stripcomments = require('strip-comments');
 const removeMarkdown = require('remove-markdown');
 
 const {defaultLocale} = require('../_data/site');
-const {isLive} = require('../_filters/live-posts');
+const {isLive} = require('../_filters/is-live');
 const {getDefaultUrl} = require('../_filters/urls');
 const {generateImgixSrc} = require('../_includes/components/Img');
 const {supportedLocales} = require('../../../shared/locale');
