+++ src/site/_collections/tags.js
@@ -27,7 +27,7 @@ const supportedTags = /** @type {{[tag: string]: unknown}} */ (
   )
 );
 
-const {isLive} = require('../_filters/live-posts');
+const {isLive} = require('../_filters/is-live');
 const {sortByUpdated} = require('../_utils/sort-by-updated');
 
 /** @type Tags */
