+++ test/unit/src/site/_filters/index.js
@@ -1,5 +1,5 @@
 describe('_filters', function () {
   require('./capitalize');
   require('./is-new-content');
-  require('./live-posts');
+  require('./is-live');
 });
